Pololu SMC02B Test.BS2
Tests for proper working of the Pololu SMC02B micro dual motor controller connected to a BS2 Ramps one motor up and then does the same for the second one and then various speeds are tested.  Code shows the use of a variable to set the speed and how to set it manually.